+++ pkg/vcdcpiclient/cpi_gateway_manager.go
