+++ pkg/vcdsdk/org.go
@@ -11,39 +11,48 @@ import (
 type OrgManager struct {
 	// client should be refreshed separately
 	Client *Client
+	OrgName string
 }
 
-func NewOrgManager(client *Client) *OrgManager{
+func NewOrgManager(client *Client, orgName string) (*OrgManager, error) {
+	// if orgName is empty, use the clusteOrgName from the client as the orgName
+	if orgName == "" && client.ClusterOrgName == ""{
+		if client.ClusterOrgName == "" {
+			return nil, fmt.Errorf("could not find a valid OrgName to create orgManager")
+		}
+		orgName = client.ClusterOrgName
+	}
 	return &OrgManager{
 		Client: client,
-	}
+		OrgName: orgName,
+	}, nil
 }
 
 func (orgManager *OrgManager) GetCatalogByName(catalogName string) (*govcd.Catalog, error) {
-	org, err := orgManager.Client.VCDClient.GetOrgByName(orgManager.Client.ClusterOrgName)
+	org, err := orgManager.Client.VCDClient.GetOrgByName(orgManager.OrgName)
 	if err != nil {
-		return nil, fmt.Errorf("unable to get vcd organization [%s]: [%v]", orgManager.Client.ClusterOrgName, err)
+		return nil, fmt.Errorf("unable to get vcd organization [%s]: [%v]", orgManager.OrgName, err)
 	}
 	if err := org.Refresh(); err != nil {
-		return nil, fmt.Errorf("unable to refresh org [%s]: [%v]", orgManager.Client.ClusterOrgName, err)
+		return nil, fmt.Errorf("unable to refresh org [%s]: [%v]", orgManager.OrgName, err)
 	}
 	catalog, err := org.GetCatalogByName(catalogName, true)
 	if err != nil {
-		return catalog, fmt.Errorf("unable to find catalog [%s] in org [%s]", catalogName, orgManager.Client.ClusterOrgName)
+		return catalog, fmt.Errorf("unable to find catalog [%s] in org [%s]", catalogName, orgManager.OrgName)
 	}
 	return catalog, nil
 }
 
 func (orgManager *OrgManager) GetComputePolicyDetailsFromName(computePolicyName string) (*types.VdcComputePolicy, error) {
-	org, err := orgManager.Client.VCDClient.GetOrgByName(orgManager.Client.ClusterOrgName)
+	org, err := orgManager.Client.VCDClient.GetOrgByName(orgManager.OrgName)
 	if err != nil {
-		return nil, fmt.Errorf("unable to get org [%s] by name: [%v]", orgManager.Client.ClusterOrgName, err)
+		return nil, fmt.Errorf("unable to get org [%s] by name: [%v]", orgManager.OrgName, err)
 	}
 
 	vdcComputePolicies, err := org.GetAllVdcComputePolicies(url.Values{})
 	if err != nil {
 		return nil, fmt.Errorf("unable to get all compute policies for [%s] by name: [%v]",
-			orgManager.Client.ClusterOrgName, err)
+			orgManager.OrgName, err)
 	}
 
 	var computePolicy *types.VdcComputePolicy = nil
